fix: harden live contract layers
