+++ package.json
@@ -77,7 +77,7 @@
     "eslint": "^10.8.0",
     "eslint-config-prettier": "^10.1.8",
     "jest": "^29.5.0",
-    "jest-junit": "^16.0.0",
+    "jest-junit": "^17.0.0",
     "prettier": "^3.9.6",
     "ts-jest": "^29.4.12",
     "ts-node": "^10.9.0",
